// 2020/Ionic/TinglewApp/src/interfaces/event-marker.interface.ts
export interface Marker {
    eventTitle: string;
    eventDescription?: string;
    eventStartDate?: string;
    eventEndDate?: string;
    eventStartTime?: string;
    eventEndTime?: string;
    eventTicketType?: string;
    eventTicketPrice?: string;
    eventRefundPolicy?: string;
    eventAddress: string;
    eventType: string;
    eventAccess: string;
    lat: string;
    lng: string;
    uid: string;
}